removed debug code
The only problem left is when other workflow is clicked/chosen, the buttons and the data wont be updated unless "refresh" button is clicked.
Will try to implement like refresh on "descriptor" from the other issue, but might not work properly cause it's a different controller.
Need to also check if button is disabled properly or not(try other github account). After workflow is done, need to also implement it on tools/containers.