fixed error and warning message
need to think of another way to either disable the publish button or to make it not publish when the error appears. It's hard to make it not change to published without changing the workflowservice.js or some of other code.
Another problem is that disabling the button wont work fully because the button/the document wont get updated unless the 'refresh' button is clicked. This also requires a bit of changing on the service(?) need to think another way to refresh it without changing the other parts of the code